game: rework a lot data model
- change data model for world / map
- change the way we draw map
- fix mistake in map aseprite (map layer)
- create legacy folder to store map import v1 code
- update the way we move player and check collisions
- remove useless graphics.rs file
- begin client.rs to simulate / brainstorm about data from futur server

+++ client/src/constants.rs
@@ -1,13 +1,13 @@
 
 // Map and Interface size
 // Add check after loading map and gui, if size differ from const stop the program
-pub const TILEMAP_WIDTH: usize = 16; 
-pub const TILEMAP_HEIGHT: usize = 12;
-pub const TILE_WIDTH: usize = 64;
-pub const TILE_HEIGHT: usize = 64;
+pub const TILEMAP_WIDTH: u32 = 16; 
+pub const TILEMAP_HEIGHT: u32 = 12;
+pub const TILE_WIDTH: u32 = 64;
+pub const TILE_HEIGHT: u32 = 64;
 
-pub const MAP_WIDTH:usize = TILE_WIDTH * TILEMAP_WIDTH;
-pub const MAP_HEIGHT:usize = TILE_HEIGHT * TILEMAP_HEIGHT;
+pub const MAP_WIDTH:usize = (TILE_WIDTH * TILEMAP_WIDTH) as usize;
+pub const MAP_HEIGHT:usize = (TILE_HEIGHT * TILEMAP_HEIGHT) as usize;
 pub const MAP_WIDTH_CENTER:usize = (WINDOW_WIDTH - MAP_WIDTH) / 2;
 pub const MAP_HEIGHT_CENTER:usize = (WINDOW_HEIGHT - GAME_HEIGHT) / 2;
